added partial keyboard controls
Mouse movement and single-clicking the primary mouse button still
have no keyboard alternatives. That should still be done.

Also, there is a bug where, before the first click in the p5js
canvas, the keyTyped event does not fire. That needs to be fixed
before this is merged to master.

Better style sheets for the help information would be a good thing
to add, particularly for <kbd> tags.