removed some unnecessary CSS for canvas
This may have been needed at one time but it doesn't appear to be
needed now. I should probably have removed it when I turned the
browser-specific CSS for avoiding undesired text selection into
stadard CSS.

Note that there is a bug, brought about by pushing down and
centering the canvas, where the top and left edges of the canvas
are at fractional-px positions, causing relative mouse cursor
positions on the canvas to be fractional. The most obvious harm of
this is that the Cartesian coordinates are fractional (and in some
browsers, at least Firefox, the y-coordinate is shown with many
digits). But it also has the problem that one cannot visualize a
ray at a right-angle. Since novices are the target audience of this
demonstration, and the sines and cosines of multiple of 90 degrees
are quite conceptually important, that may actually be the more
serious effect. This definitely needs to be fixed.

That bug exists with or without the CSS I am removing here, which
doesn't appear to affect it. This commit doesn't introduce the bug;
it was introduced in commit adb5c9e since that's where the body was
horizontally centered (and padded) and where elements that appear
above the canvas (an <h1> and a <p>) were added. I'm only
mentioning it here so that if I don't manage to fix it soon and
look back at this commit later, I don't come to wrongly believe
that this commit caused, alleviated, or was intended to address it.