fixed fractional pixels bug
The bug this fixes is described in the commit message for c24b3f3.

Since the problem comes down to how integer-valued
viewport-relative mouse cursor positions are translated into
sometimes fractional canvas-relative positions, I've just fixed the
translation by rounding the computed canvas-relative positions
before using them.

Since the input canvas coordinates are rounded before being used,
rather than taking the shallow alternative of rounding the output
display coordinates before showing them, this approach doesn't
merely fix the problem of wrong displayed coordinates, but also the
problem that the bug usually prevented the ray from being rotated
to coincide exactly with an axis.

It's possible that the rounding I'm using is different from the
rounding scheme (some?) browsers use, in which case the pointer may
be a pixel off due to rounding. But this error should be consistent
and isotropic, so it should not cause any problems (and is unlikely
to be noticed).

If I make this program work in mobile browsers--where tapping the
canvas will move the dot to the tapped location--then I should keep
this situation in mind to avoid introducing a regression.